UserServiceImplTest: adapted tests after changing countByLogin() return type.
Correction for b2ef0421bf37e92872fb81b70019915391d341f6 commit.

No functional changes.

+++ src/test/groovy/ru/mystamps/web/service/UserServiceImplTest.groovy
@@ -347,11 +347,11 @@ class UserServiceImplTest extends Specification {
 	
 	def "countByLogin() should call dao"() {
 		given:
-			jdbcUserDao.countByLogin(_ as String) >> 2
+			jdbcUserDao.countByLogin(_ as String) >> 2L
 		when:
-			int result = service.countByLogin('any-login')
+			long result = service.countByLogin('any-login')
 		then:
-			result == 2
+			result == 2L
 	}
 	
 	def "countByLogin() should pass login to dao"() {
